COMMERCE-5205 Moved configuration in another package
