+++ modules/apps/commerce/commerce-inventory-service/src/main/java/com/liferay/commerce/inventory/internal/messaging/CheckCommerceInventoryBookedQuantityMessageListener.java
@@ -14,7 +14,7 @@
 
 package com.liferay.commerce.inventory.internal.messaging;
 
-import com.liferay.commerce.inventory.internal.configuration.CommerceInventorySystemConfiguration;
+import com.liferay.commerce.inventory.configuration.CommerceInventorySystemConfiguration;
 import com.liferay.commerce.inventory.service.CommerceInventoryBookedQuantityLocalService;
 import com.liferay.portal.configuration.metatype.bnd.util.ConfigurableUtil;
 import com.liferay.portal.kernel.messaging.BaseMessageListener;
