+++ portal-impl/src/portal-osgi-configuration.properties
@@ -1173,15 +1173,15 @@
     #
     # Env: LIFERAY_CONFIGURATION_PERIOD_OVERRIDE_PERIOD_COM_PERIOD_LIFERAY_PERIOD_COMMERCE_PERIOD_INVENTORY_PERIOD_INTERNAL_PERIOD_CONFIGURATION_PERIOD__UPPERCASEC_OMMERCE_UPPERCASEI_NVENTORY_UPPERCASES_YSTEM_UPPERCASEC_ONFIGURATION_UNDERLINE_CHECK_UPPERCASEC_OMMERCE_UPPERCASEI_NVENTORY_UPPERCASET_EMPORARY_UPPERCASEB_OOKED_UPPERCASEQ_UANTITY_UPPERCASEI_NTERVAL
     #
-    #configuration.override.com.liferay.commerce.inventory.internal.configuration.CommerceInventorySystemConfiguration_checkCommerceInventoryTemporaryBookedQuantityInterval=
+    #configuration.override.com.liferay.commerce.inventory.configuration.CommerceInventorySystemConfiguration_checkCommerceInventoryTemporaryBookedQuantityInterval=
     #
     # Env: LIFERAY_CONFIGURATION_PERIOD_OVERRIDE_PERIOD_COM_PERIOD_LIFERAY_PERIOD_COMMERCE_PERIOD_INVENTORY_PERIOD_INTERNAL_PERIOD_CONFIGURATION_PERIOD__UPPERCASEC_OMMERCE_UPPERCASEI_NVENTORY_UPPERCASES_YSTEM_UPPERCASEC_ONFIGURATION_UNDERLINE_CHECK_UPPERCASEC_OMMERCE_UPPERCASEI_NVENTORY_UPPERCASEA_UDIT_UPPERCASEQ_UANTITY_UPPERCASEI_NTERVAL
     #
-    #configuration.override.com.liferay.commerce.inventory.internal.configuration.CommerceInventorySystemConfiguration_checkCommerceInventoryAuditQuantityInterval=
+    #configuration.override.com.liferay.commerce.inventory.configuration.CommerceInventorySystemConfiguration_checkCommerceInventoryAuditQuantityInterval=
     #
     # Env: LIFERAY_CONFIGURATION_PERIOD_OVERRIDE_PERIOD_COM_PERIOD_LIFERAY_PERIOD_COMMERCE_PERIOD_INVENTORY_PERIOD_INTERNAL_PERIOD_CONFIGURATION_PERIOD__UPPERCASEC_OMMERCE_UPPERCASEI_NVENTORY_UPPERCASES_YSTEM_UPPERCASEC_ONFIGURATION_UNDERLINE_DELETE_UPPERCASEA_UDIT_UPPERCASEM_ONTH_UPPERCASEI_NTERVAL
     #
-    #configuration.override.com.liferay.commerce.inventory.internal.configuration.CommerceInventorySystemConfiguration_deleteAuditMonthInterval=
+    #configuration.override.com.liferay.commerce.inventory.configuration.CommerceInventorySystemConfiguration_deleteAuditMonthInterval=
     #
     # Env: LIFERAY_CONFIGURATION_PERIOD_OVERRIDE_PERIOD_COM_PERIOD_LIFERAY_PERIOD_COMMERCE_PERIOD_MEDIA_PERIOD_INTERNAL_PERIOD_CONFIGURATION_PERIOD__UPPERCASEC_OMMERCE_UPPERCASEM_EDIA_UPPERCASED_EFAULT_UPPERCASEI_MAGE_UPPERCASEC_ONFIGURATION_UNDERLINE_DEFAULT_UPPERCASEF_ILE_UPPERCASEE_NTRY_UPPERCASEI_D
     #
